# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError generic error
    type: object
    x-go-package: github.com/grkmk/glm-api/models
  Product:
    description: Product Product defines the sturcture for an API product
    properties:
      description:
        description: description
        type: string
        x-go-name: Description
      id:
        description: the id for this user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: name
        type: string
        x-go-name: Name
      price:
        description: price
        format: float
        type: number
        x-go-name: Price
      sku:
        description: s k u
        type: string
        x-go-name: SKU
    required:
    - id
    type: object
    x-go-package: github.com/grkmk/glm-api/models
  ValidationError:
    description: ValidationError validation error
    type: object
    x-go-package: github.com/grkmk/glm-api/models
info:
  description: Documentation for Product API
  title: Product API.
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      parameters:
      - description: The currency used when returning the price of the product
        in: query
        name: Currency
        type: string
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Create new product
      operationId: addProduct
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product from the store
      operationId: deleteProduct
      parameters:
      - description: The id of the product to delete from the data store
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
    get:
      description: Returns a product
      operationId: listProduct
      parameters:
      - description: The currency used when returning the price of the product
        in: query
        name: Currency
        type: string
      responses:
        "200":
          $ref: '#/responses/productResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContent:
    description: ""
  productResponse:
    description: A product returns in the response
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products returns in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"